Fix end address for assertion

+++ embassy-stm32/src/flash/common.rs
@@ -91,7 +91,7 @@ pub(super) fn blocking_read(base: u32, size: u32, offset: u32, bytes: &mut [u8])
     let start_address = base + offset;
 
     #[cfg(flash_f4)]
-    family::assert_not_corrupted_read(start_address + bytes.len());
+    family::assert_not_corrupted_read(start_address + bytes.len() as u32);
 
     let flash_data = unsafe { core::slice::from_raw_parts(start_address as *const u8, bytes.len()) };
     bytes.copy_from_slice(flash_data);
